requirements.txt
# AI Qube Centaur System - Python Requirements

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI Integration
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2

# Vector Database & RAG
pinecone-client==2.2.4
weaviate-client==3.25.3
chromadb==0.4.18
sentence-transformers==2.2.2

# Data Processing
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# Database & Caching
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.0

# Async Processing
celery==5.3.4
kombu==5.3.4

# HTTP & Web
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Configuration & Environment
python-dotenv==1.0.0
pydantic-settings==2.1.0
dynaconf==3.2.4

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==41.0.8

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx[test]==0.25.2

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Workflow & Process Management
workflow-engine==0.1.0
state-machine==0.2.0

# Machine Learning & AI
torch==2.1.1
transformers==4.36.0
huggingface-hub==0.19.4

# Data Validation
marshmallow==3.20.1
cerberus==1.3.5

# Task Queue
rq==1.15.1
dramatiq==1.14.2

# Time & Date
python-dateutil==2.8.2
arrow==1.3.0

# Utilities
click==8.1.7
rich==13.7.0
typer==0.9.0
tqdm==4.66.1

# Network & Communication
websockets==12.0
socketio==0.2.1

# File Processing
PyYAML==6.0.1
toml==0.10.2
xmltodict==0.13.0

# Performance Monitoring
psutil==5.9.6
memory-profiler==0.61.0

# Recursive Learning & Meta-Learning
stable-baselines3==2.2.1
gymnasium==0.29.1

# Pattern Recognition
pattern-recognition==0.1.0
similarity-engine==0.2.0
